Fix 64bit issues
The driver heavily uses "unsigned int" to store pointers. However, this
isn't safe on 64bit machines to change this to "unsigned long" which
should always work without any issues.

This also fixes some printf() specifier issues which are wrongly used.

+++ core/rtw_br_ext.c
@@ -126,7 +126,7 @@ static int skb_pull_and_merge(struct sk_buff *skb, unsigned char *src, int len)
 	int tail_len;
 	unsigned long end, tail;
 
-	if ((src+len) > skb->tail || skb->len < len)
+	if ((unsigned long)(src+len) > skb->tail || skb->len < len)
 		return -1;
 
 	tail = (unsigned long)skb->tail;
@@ -1633,13 +1633,13 @@ void dhcp_flag_bcast(_adapter *priv, struct sk_buff *skb)
 
 			if(iph->protocol == IPPROTO_UDP) // UDP
 			{
-				struct udphdr *udph = (struct udphdr *)((unsigned int)iph + (iph->ihl << 2));
+				struct udphdr *udph = (struct udphdr *)((unsigned long)iph + (iph->ihl << 2));
 
 				if((udph->source == __constant_htons(CLIENT_PORT))
 					&& (udph->dest == __constant_htons(SERVER_PORT))) // DHCP request
 				{
 					struct dhcpMessage *dhcph =
-						(struct dhcpMessage *)((unsigned int)udph + sizeof(struct udphdr));
+						(struct dhcpMessage *)((unsigned long)udph + sizeof(struct udphdr));
 
 					if(dhcph->cookie == __constant_htonl(DHCP_MAGIC)) // match magic word
 					{
